7699 move CachedScopeMixin
